requirements.txt
numpy
pillow
transformers
torch
fastapi
python-multipart
uvicorn